// src/store/user/user.types.ts
export enum UserActionTypes {
  checkSession = 'checkSession',
  googleSignInStart = 'googleSignInStart',
  emailSignInStart = 'emailSignInStart',
  signInSuccess = 'signInSuccess',
  signInFailed = 'signInFailed',
  signUpStart = 'signUpStart',
  signUpSuccess = 'signUpSuccess',
  signUpFailed = 'signUpFailed',
  signOutStart = 'signOutStart',
  signOutSuccess = 'signOutSuccess',
  signOutFailed = 'signOutFailed',
}